Move `pages` to fixtures file
